toolkit/spatial-navigation/tests/chrome/test_snav_disabledElement.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css type="text/css"?>

<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=436084
-->

<window title="Mozilla Bug 288254"
  xmlns:html="http://www.w3.org/1999/xhtml"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
  onload="onLoad();">

  <script type="application/javascript" src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>      

<body id="some-content" xmlns="http://www.w3.org/1999/xhtml">

<p>
<a id="start" href="https://bugzilla.mozilla.org/show_bug.cgi?id=436084">Mozilla Bug 436084</a>
</p>

<p>
<input id="hi_mom" value="Some text" disabled="true" type="text"> </input>
</p>

<p>
<a id="end" href="https://bugzilla.mozilla.org/show_bug.cgi?id=436084">Mozilla Bug 436084</a>
</p>

</body>


<script class="testbody" type="application/javascript">
<![CDATA[
Components.utils.import("resource://gre/modules/SpatialNavigation.js");

function onLoad()
{
    var x = document.getElementById("some-content");
    var snav = new SpatialNavigation(x);

    function moveAndVerify(direction, value)
    {
        sendKey(direction, document.activeElement);
        ok(document.activeElement.getAttribute("id") == value, "Move");
    }
    
    // get to a known place.
    document.getElementById("start").focus();
    
    // from start.
    moveAndVerify("DOWN", "end");
    
    document.getElementById("hi_mom").disabled=false; 
    
    moveAndVerify("UP", "hi_mom");
    
    SimpleTest.finish();
}

]]></script>
</window>
